+++ src/components/icons/BackArrow.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const BackArrow = ({ width, height, color }: IconProps) => {
   return (
