+++ src/components/icons/Calendar.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const Calendar = ({ width, height, color }: IconProps) => {
   return (
