+++ src/components/icons/Close.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const Close = ({ width, height, color }: IconProps) => {
   return (
