+++ src/components/icons/DownArrow.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const DownArrow = ({ width, height, color }: IconProps) => {
   return (
