+++ src/components/icons/DownwardArrow.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const DownwardArrow = ({ width, height, color }: IconProps) => {
   return (
