+++ src/components/icons/Email.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const Email = ({ width, height }: IconProps) => {
   return (
