+++ src/components/icons/Setting.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const Setting = ({ width, height, color }: IconProps) => {
   return (
