+++ src/components/icons/Speech.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const Speech = ({ width, height, color }: IconProps) => {
   return (
