+++ src/components/icons/UpArrow.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const UpArrow = ({ width, height, color }: IconProps) => {
   return (
