+++ src/components/icons/User.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from "../../interfaces/iconProps";
+import { IconProps } from "../../types/iconProps";
 
 export const User = ({ width, height, color }: IconProps) => {
   return (
