+++ src/components/organisms/Profile.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { userInterface } from "../../interfaces/interface";
+import { userInterface } from "../../types/interfaces";
 
 interface ProfileProps {
   nonTypingUser: userInterface;
