+++ src/components/organisms/SingleCard.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 // styles
 import styled from "styled-components";
 // interface
-import { userInterface, chattingInterface } from "../../interfaces/interface";
+import { userInterface, chattingInterface } from "../../types/interfaces";
 // json
 import usersData from "../../json/usersData.json";
 import { getDateString } from "../../utils/getDateString";
