--- src/hooks/useClickedCoord.tsx
@@ -1,14 +1,0 @@
-import { useState } from "react";
-
-const useClickedCoord = () => {
-  // const [coords, setCoords] = useState({ x: 0, y: 0 });
-
-  // const handleClick = (e: React.MouseEvent<HTMLDivElement>) => {
-  //   setIsSorting(!isSorting);
-  //   setCoords({ x: e.clientX, y: e.clientY });
-  // };
-
-  return 1;
-};
-
-export default useClickedCoord;
